[FEAT]로그인,회원가입 화면 구현 및 메인 화면 디자인 수정
