+++ src/components/Layout/Layout.jsx
@@ -1,12 +1,12 @@
 import React from "react";
-import Sidebar from "../components/Sidebar/Sidebar";
-import Main from "../components/Main/Main";
+import Sidebar from "../Sidebar/Sidebar";
+import { Outlet } from "react-router-dom";
 
-export default function HomePage({ onLogout }) {
+export default function Layout({ onLogout }) {
   return (
     <div style={{ display: "flex", height: "100vh", width: "100%" }}>
       <Sidebar onLogout={onLogout} />
-      <Main />
+      <Outlet />
     </div>
   );
 }
