+++ src/pages/LoginPage.css
@@ -19,6 +19,8 @@
   --input-bg: #f3f4f6;
   --input-border: #e5e7eb;
   --shadow: 0 10px 30px rgba(0,0,0,0.18);
+  --danger: #d93025;
+  --success: #1b8a3d;
 }
 
 html, body{
@@ -77,7 +79,7 @@ body{
   border-radius: 10px;
   font-weight: 700;
   letter-spacing: 0.2px;
-  margin-top: 12px;
+  margin-top: 15px;
   cursor: pointer;
   box-shadow: var(--shadow);
   transition: transform .15s ease, box-shadow .15s ease, opacity .15s ease, filter .15s ease;
@@ -105,17 +107,17 @@ body{
   flex-direction: column;
   padding: 0 40px;
   height: 100%;
-  gap: 10px;
+  gap: 2px; /* 회원가입 폼 간격 조정 */
 }
 
 .container input{
   background-color: var(--input-bg);
   border: none;
-  margin: 8px 0;
-  padding: 14px 14px;
-  font-size: 15px;
-  border-radius: 10px;
-  width: 60%;
+  margin: 4px 0; /* 회원가입 폼 간격 조정 */
+  padding: 12px 14px; /* 회원가입 폼 패딩 조정 */
+  font-size: 14px; /* 회원가입 폼 폰트 크기 조정 */
+  border-radius: 8px; /* 회원가입 폼 테두리 조정 */
+  width: 100%;
   outline: none;
   transition: border-color .15s ease, box-shadow .15s ease;
 }
@@ -131,7 +133,22 @@ body{
   top: 0;
   height: 100%;
   transition: all 0.6s ease-in-out;
+  /* 회원가입 폼이 길어지면 스크롤 */
+  overflow-y: auto;
 }
+
+/* 스크롤바 스타일 (선택적) */
+.form-container::-webkit-scrollbar {
+  width: 6px;
+}
+.form-container::-webkit-scrollbar-thumb {
+  background-color: var(--bg2);
+  border-radius: 3px;
+}
+.form-container::-webkit-scrollbar-track {
+  background-color: var(--input-bg);
+}
+
 
 .sign-in{
   left: 0;
@@ -164,23 +181,124 @@ body{
   display: flex;
   gap: 8px;
 }
-.social-icons a{
-  border: 1px solid #e5e7eb;
-  background: #fff;
-  border-radius: 10px;
-  display: inline-flex;
-  justify-content: center;
+
+.sign-in form input {
+  width: 70%; /* 로그인 폼 입력 필드 너비 증가 */
+  padding: 14px 14px; /* 로그인 폼 입력 필드 패딩 증가 */
+  font-size: 16px; /* 로그인 폼 입력 필드 폰트 크기 증가 */
+}
+
+.sign-in form input:not(:last-of-type) {
+  margin-bottom: 25px; /* 로그인 폼 아이디와 비밀번호 사이 간격 */
+}
+
+/* --- 회원가입 폼 상세 스타일 --- */
+.sign-up .form-group {
+  width: 70%;
+  margin-bottom: 2px;
+  position: relative;
+}
+
+.sign-up .id-check-row, .sign-up .address-group {
+  display: flex;
+  gap: 8px;
   align-items: center;
-  width: 44px;
-  height: 44px;
-  color: #1f2937;
-  transition: background .15s ease, transform .15s ease, box-shadow .15s ease;
 }
-.social-icons a:hover{
-  background: #f8fafc;
-  transform: translateY(-1px);
-  box-shadow: 0 6px 16px rgba(0,0,0,0.08);
+
+.sign-up .id-check-row {
+  width: 100%;
 }
+
+.sign-up .id-check-row input, .sign-up .address-group input {
+  flex: 1;
+  width: auto; /* flex 아이템이므로 width 초기화 */
+}
+
+.sign-up .id-check-row button, .sign-up .address-group button {
+  padding: 10px 14px;
+  font-size: 13px;
+  margin-top: 0;
+  width: auto; /* 자동 너비 */
+  flex-shrink: 0; /* 줄어들지 않도록 */
+}
+
+.id-check-message {
+  font-size: 12px;
+  color: var(--primary);
+  width: 100%;
+  text-align: left;
+  margin-top: 2px;
+  height: 16px; /* 메시지 영역 확보 */
+}
+.id-check-message:not(:empty) {
+  margin-bottom: 4px;
+}
+
+.password-group {
+  width: 100%;
+  position: relative;
+}
+.password-group .inline-warning {
+  font-size: 12px;
+  color: var(--danger);
+  text-align: left;
+  padding-left: 2px;
+  height: 18px; /* 메시지가 나타날 때 차지할 높이 */
+  line-height: 18px; /* 텍스트를 수직 중앙에 맞춤 */
+}
+
+.form-group.agreements {
+  text-align: left;
+}
+
+.agreement-item {
+  display: flex;
+  align-items: center;
+  margin: 6px 0;
+  font-size: 13px;
+  color: var(--muted);
+}
+.agreement-item input {
+  margin-right: 8px;
+  width: auto; /* 체크박스 크기 초기화 */
+}
+.agreement-item span {
+  color: var(--muted);
+}
+
+/* 메시지 스타일 */
+.error-message, .success-message {
+  width: 60%;
+  border-radius: 8px;
+  padding: 10px;
+  text-align: center;
+  margin: 5px 0;
+  font-size: 13px;
+  font-weight: 500;
+}
+
+.error-message {
+  color: var(--danger);
+  background-color: #fdecea;
+}
+
+.success-message {
+  color: var(--success);
+  background-color: #e6f4ea;
+}
+
+.signup-link {
+  font-size: 13px !important;
+}
+.signup-link span {
+  color: var(--primary) !important;
+  cursor: pointer;
+  font-weight: 600;
+}
+.signup-link span:hover {
+  text-decoration: underline;
+}
+
 
 /* 토글 패널 */
 .toggle-container{
@@ -228,7 +346,12 @@ body{
 
 /* 반응형 */
 @media (max-width: 960px){
-  .container input { width: 72%; }
+  .container input, .error-message, .success-message {
+    width: 72%;
+  }
+  .sign-up .form-group {
+    width: 80%;
+  }
 }
 @media (max-width: 820px){
   .toggle-container{ display: none; }
@@ -236,5 +359,7 @@ body{
   .container.active .sign-in,
   .container.active .sign-up{ transform: none; }
   .container form{ padding: 0 24px; }
-  .container input{ width: 100%; }
+  .container input, .error-message, .success-message, .sign-up .form-group {
+    width: 100%;
+  }
 }
