--- src/pages/SignupPage.css
@@ -1,206 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Outfit:wght@100..900&display=swap');
-
-html, body, #root {
-  height: 100%;
-  margin: 0;
-  padding: 0;
-}
-
-* {
-  box-sizing: border-box;
-  font-family: "Outfit", sans-serif;
-}
-.signup-page {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    min-height: 100vh;
-    width: 100%;
-    background-color: #f5f5f5;
-    animation: fadeIn 1.2s ease-in-out;
-  
-    /* 🔥 추가: 폼이 화면보다 클 때 스크롤 가능하도록 */
-    overflow-y: auto;
-    padding: 40px 0; /* 위아래 여백 추가로 중앙 정렬 유지 */
-  }
-  
-  .signup-form {
-    width: 360px;
-    background: white;
-    padding: 40px;
-    border-radius: 12px;
-    box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
-  
-    /* 🔥 중앙에서 튀어나오지 않게 max-height 제한 */
-    max-height: 100%;
-    overflow-y: auto;
-  }
-  
-
-.signup-form h2 {
-  text-align: center;
-  color: #007bff;
-  margin-bottom: 25px;
-}
-
-/* form layout */
-.form-group {
-  margin-bottom: 15px;
-  position: relative;
-  min-height: 70px; /* 🔥 메시지 생겨도 밀리지 않도록 고정 높이 */
-}
-
-.form-group label {
-  display: block;
-  margin-bottom: 5px;
-  font-weight: 500;
-}
-
-.form-group input {
-  width: 100%;
-  padding: 10px;
-  border: 1px solid #ccc;
-  border-radius: 5px;
-}
-
-.form-group input:focus {
-  border-color: #007bff;
-  outline: none;
-}
-
-/* 아이디 중복확인 */
-.id-check-row {
-  display: flex;
-  gap: 8px;
-}
-
-.id-check-row input {
-  flex: 1;
-}
-
-.id-check-row button {
-  padding: 8px 12px;
-  border: none;
-  background-color: #007bff;
-  color: white;
-  border-radius: 5px;
-  cursor: pointer;
-  font-weight: 500;
-}
-
-.id-check-row button:hover {
-  background-color: #0069d9;
-}
-
-.id-check-message {
-  margin-top: 5px;
-  font-size: 0.85rem;
-  color: #007bff;
-}
-
-/* 비밀번호 불일치 경고 */
-.password-container {
-  position: relative;
-}
-
-.inline-warning {
-  position: absolute;
-  bottom: -18px;
-  left: 2px;
-  color: #d93025;
-  font-size: 0.8rem;
-}
-.inline-good {
-    position: absolute;
-    bottom: -18px;
-    left: 2px;
-    color: #25d943;
-    font-size: 0.8rem;
-  }
-
-/* 약관 영역 */
-.agreements {
-  margin-top: 15px;
-  margin-bottom: 10px;
-}
-
-.agreement-item {
-  display: flex;
-  align-items: center;
-  margin: 4px 0;
-  font-size: 0.9rem;
-  color: #333;
-}
-
-.agreement-item input {
-  margin-right: 6px;
-}
-
-/* 버튼 */
-button[type="submit"] {
-  width: 100%;
-  background-color: #007bff;
-  color: white;
-  padding: 10px;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-  transition: 0.2s ease;
-  font-weight: 600;
-}
-
-button[type="submit"]:hover {
-  background-color: #0069d9;
-}
-
-/* 메시지 */
-.error-message {
-  color: #d93025;
-  background-color: #fdecea;
-  border: 1px solid #f5c6cb;
-  border-radius: 5px;
-  padding: 8px;
-  text-align: center;
-  margin-bottom: 10px;
-  font-size: 0.9rem;
-}
-
-.success-message {
-  color: #1b8a3d;
-  background-color: #e6f4ea;
-  border: 1px solid #a3d6aa;
-  border-radius: 5px;
-  padding: 8px;
-  text-align: center;
-  margin-bottom: 10px;
-  font-size: 0.9rem;
-}
-
-/* 로그인으로 이동 */
-.login-link {
-  text-align: center;
-  margin-top: 1rem;
-  color: #555;
-  font-size: 0.9rem;
-}
-
-.login-link span {
-  color: #007bff;
-  cursor: pointer;
-  font-weight: bold;
-}
-
-.login-link span:hover {
-  text-decoration: underline;
-}
-
-@keyframes fadeIn {
-  0% {
-    opacity: 0;
-    transform: translateY(10px);
-  }
-  100% {
-    opacity: 1;
-    transform: translateY(0);
-  }
-}
